+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/mongo/MongoClientFactory.java
@@ -16,21 +16,15 @@
 
 package org.springframework.boot.autoconfigure.mongo;
 
-import java.util.Collections;
 import java.util.List;
 
-import com.mongodb.ConnectionString;
-import com.mongodb.MongoClientSettings;
-import com.mongodb.MongoClientSettings.Builder;
-import com.mongodb.MongoCredential;
-import com.mongodb.ServerAddress;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 
 import org.springframework.core.env.Environment;
 
 /**
- * A factory for a blocking {@link MongoClient} that applies {@link MongoProperties}.
+ * A factory for a blocking {@link MongoClient}.
  *
  * @author Dave Syer
  * @author Phillip Webb
@@ -40,90 +34,32 @@ import org.springframework.core.env.Environment;
  * @author Stephane Nicoll
  * @author Nasko Vasilev
  * @author Mark Paluch
+ * @author Scott Frederick
  * @since 2.0.0
  */
-public class MongoClientFactory {
+public class MongoClientFactory extends MongoClientFactorySupport<MongoClient> {
 
-	private final MongoProperties properties;
-
-	private final Environment environment;
-
+	/**
+	 * Construct a factory for creating a blocking {@link MongoClient}.
+	 * @param properties configuration properties
+	 * @param environment a Spring {@link Environment} containing configuration properties
+	 * @deprecated since 2.3.0 in favor of
+	 * {@link #MongoClientFactory(MongoProperties, Environment, List)}
+	 */
+	@Deprecated
 	public MongoClientFactory(MongoProperties properties, Environment environment) {
-		this.properties = properties;
-		this.environment = environment;
+		this(properties, environment, null);
 	}
 
 	/**
-	 * Creates a {@link MongoClient} using the given {@link MongoClientSettings settings}.
-	 * If the environment contains a {@code local.mongo.port} property, it is used to
-	 * configure a client to an embedded MongoDB instance.
-	 * @param settings the settings
-	 * @return the Mongo client
+	 * Construct a factory for creating a blocking {@link MongoClient}.
+	 * @param properties configuration properties
+	 * @param environment a Spring {@link Environment} containing configuration properties
+	 * @param builderCustomizers a list of configuration settings customizers
 	 */
-	public MongoClient createMongoClient(MongoClientSettings settings) {
-		Builder settingsBuilder = (settings != null) ? MongoClientSettings.builder(settings)
-				: MongoClientSettings.builder();
-		settingsBuilder.uuidRepresentation(this.properties.getUuidRepresentation());
-		Integer embeddedPort = getEmbeddedPort();
-		if (embeddedPort != null) {
-			return createEmbeddedMongoClient(settingsBuilder, embeddedPort);
-		}
-		return createNetworkMongoClient(settingsBuilder);
-	}
-
-	private Integer getEmbeddedPort() {
-		if (this.environment != null) {
-			String localPort = this.environment.getProperty("local.mongo.port");
-			if (localPort != null) {
-				return Integer.valueOf(localPort);
-			}
-		}
-		return null;
-	}
-
-	private MongoClient createEmbeddedMongoClient(Builder settings, int port) {
-		String host = (this.properties.getHost() != null) ? this.properties.getHost() : "localhost";
-		settings.applyToClusterSettings(
-				(cluster) -> cluster.hosts(Collections.singletonList(new ServerAddress(host, port))));
-		return MongoClients.create(settings.build());
-	}
-
-	private MongoClient createNetworkMongoClient(Builder settings) {
-		MongoProperties properties = this.properties;
-		if (properties.getUri() != null) {
-			return createMongoClient(properties.getUri(), settings);
-		}
-		if (hasCustomAddress() || hasCustomCredentials()) {
-			if (hasCustomCredentials()) {
-				String database = (this.properties.getAuthenticationDatabase() != null)
-						? this.properties.getAuthenticationDatabase() : this.properties.getMongoClientDatabase();
-				settings.credential((MongoCredential.createCredential(this.properties.getUsername(), database,
-						this.properties.getPassword())));
-			}
-			String host = getValue(properties.getHost(), "localhost");
-			int port = getValue(properties.getPort(), MongoProperties.DEFAULT_PORT);
-			List<ServerAddress> seeds = Collections.singletonList(new ServerAddress(host, port));
-			settings.applyToClusterSettings((cluster) -> cluster.hosts(seeds));
-			return MongoClients.create(settings.build());
-		}
-		return createMongoClient(MongoProperties.DEFAULT_URI, settings);
-	}
-
-	private MongoClient createMongoClient(String uri, Builder settings) {
-		settings.applyConnectionString(new ConnectionString(uri));
-		return MongoClients.create(settings.build());
-	}
-
-	private <T> T getValue(T value, T fallback) {
-		return (value != null) ? value : fallback;
-	}
-
-	private boolean hasCustomAddress() {
-		return this.properties.getHost() != null || this.properties.getPort() != null;
-	}
-
-	private boolean hasCustomCredentials() {
-		return this.properties.getUsername() != null && this.properties.getPassword() != null;
+	public MongoClientFactory(MongoProperties properties, Environment environment,
+			List<MongoClientSettingsBuilderCustomizer> builderCustomizers) {
+		super(properties, environment, builderCustomizers, MongoClients::create);
 	}
 
 }
